nil ratings no longer report as F (#30)
Fixes #15

Previously, when ratings were missing, it would be reported as an F/0
With the repair averages, there was a calculation issue where if there
were no entries, or all the entries were nil, it would also report as
F/0. This commit adds some guard clauses and pre-checks to ensure that
the value is calculated accurately.